Tests1/Services/HackerNewsServiceTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using HackerNews.Application.Services;
using Moq;
using Xunit;
using Microsoft.Extensions.Caching.Memory;
using System.Net.Http.Json;
using HackerNews.Application.Contracts.Models;
using Autofac.Extras.Moq;
using Newtonsoft.Json;
using System.Net;
using Moq.Protected;

namespace HackerNews.Application.Services.Tests
{
    [Trait("Service", nameof(HackerNewsService))]
    [TestClass()]
    public class HackerNewsServiceTests
    {
        [TestMethod]
        [Fact]
        public async Task GetNewestStoriesAsync_ReturnsStoriesSuccessfully()
        {
            var hackerNewsApiUrl = "https://hacker-news.firebaseio.com/v0/";
            // Arrange
            var hackerNewsService = new HackerNewsService(hackerNewsApiUrl); // Assuming the class containing the methods is named HackerNewsService

            // Act
            var result = await hackerNewsService.GetNewestStoriesAsync();

            // Assert
            Xunit.Assert.NotNull(result);
            Xunit.Assert.True(result.Count > 0);
            // Add more assertions based on your specific requirements
        }

        [TestMethod]
        [Fact]
        public async Task FetchStoryAsync_ReturnsStorySuccessfully()
        {
            var hackerNewsApiUrl = "https://hacker-news.firebaseio.com/v0/";
            // Arrange
            var hackerNewsService = new HackerNewsService(hackerNewsApiUrl); // Assuming the class containing the methods is named HackerNewsService
            var httpClient = new HttpClient();

            // Act
            var result = await hackerNewsService.FetchStoryAsync(httpClient, "sampleStoryUrl", new MemoryCache(new MemoryCacheOptions()), 123);

            // Assert
            Xunit.Assert.NotNull(result);
            // Add more assertions based on your specific requirements
        }

        [TestMethod]
        [Fact]
        public async Task FetchStoryAsync_HandlesExceptionGracefully()
        {
            var hackerNewsApiUrl = "https://hacker-news.firebaseio.com/v0/";
            // Arrange
            var hackerNewsService = new HackerNewsService(hackerNewsApiUrl); // Assuming the class containing the methods is named HackerNewsService
            var httpClient = new HttpClient();

            // Act
            var result = await hackerNewsService.FetchStoryAsync(httpClient, "invalidUrl", new MemoryCache(new MemoryCacheOptions()), 123);

            // Assert
            Xunit.Assert.NotNull(result);
            // Add more assertions based on your specific requirements for handling exceptions
        }
    }
}